RYA-489 Moved common Accumulo integration test code to the rya.test.accumulo project.
