+++ dao/accumulo.rya/pom.xml
@@ -61,7 +61,8 @@ under the License.
         <dependency>
             <groupId>commons-io</groupId>
             <artifactId>commons-io</artifactId>
-            </dependency>
+        </dependency>
+            
         <!--  testing dependencies -->
         <dependency>
             <groupId>org.eclipse.rdf4j</groupId>
@@ -80,29 +81,13 @@ under the License.
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.apache.accumulo</groupId>
-            <artifactId>accumulo-minicluster</artifactId>
+            <groupId>org.apache.rya</groupId>
+            <artifactId>rya.test.accumulo</artifactId>
             <scope>test</scope>
         </dependency>
     </dependencies>
 
     <build>
-        <plugins>
-            <plugin>
-                <!--  generate the test jar as well so it can be reused by dependent tools.
-                TODO this is messy.  in the future, classes that provide this functionality 
-                should be decoupled into reusable frameworks. -->
-                <artifactId>maven-jar-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>test-jar</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    
         <pluginManagement>
             <plugins>
                 <plugin>
@@ -133,7 +118,6 @@ under the License.
                                 </configuration>
                             </execution>
                         </executions>
-
                     </plugin>
                 </plugins>
             </build>
