+++ extras/indexing/src/test/java/org/apache/rya/api/client/accumulo/AccumuloListInstancesIT.java
@@ -28,9 +28,9 @@ import org.apache.accumulo.core.client.AccumuloSecurityException;
 import org.apache.accumulo.core.client.admin.SecurityOperations;
 import org.apache.accumulo.core.client.admin.TableOperations;
 import org.apache.accumulo.core.security.SystemPermission;
-import org.apache.rya.accumulo.AccumuloITBase;
 import org.apache.rya.api.client.Install;
 import org.apache.rya.api.client.Install.InstallConfiguration;
+import org.apache.rya.test.accumulo.AccumuloITBase;
 import org.apache.rya.api.client.ListInstances;
 import org.apache.rya.api.client.RyaClientException;
 import org.junit.Before;
