+++ extras/shell/pom.xml
@@ -88,25 +88,19 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.apache.accumulo</groupId>
-            <artifactId>accumulo-minicluster</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
             <groupId>org.apache.fluo</groupId>
             <artifactId>fluo-mini</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.apache.rya</groupId>
-            <artifactId>accumulo.rya</artifactId>
+            <artifactId>mongodb.rya</artifactId>
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.apache.rya</groupId>
-            <artifactId>mongodb.rya</artifactId>
-            <type>test-jar</type>
+            <artifactId>rya.test.accumulo</artifactId>
             <scope>test</scope>
         </dependency>
     </dependencies>
