+++ test/accumulo/src/main/java/org/apache/rya/test/accumulo/MiniAccumuloClusterInstance.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.rya.accumulo;
+package org.apache.rya.test.accumulo;
 
 import java.io.File;
 import java.io.IOException;
