+++ test/pom.xml
@@ -34,6 +34,7 @@
     <packaging>pom</packaging>
 
     <modules>
+        <module>accumulo</module>
         <module>kafka</module>
         <module>rdf</module>
     </modules>
